+++ internal/convert/convert.go
@@ -206,10 +206,11 @@ func fromPbSubscriptionConfig(config *pb.SubscriptionConfig) primitive.Subscript
 		return primitive.SubscriptionConfig{}
 	}
 	to := primitive.SubscriptionConfig{
-		RateLimit:        config.RateLimit,
-		MaxRetryAttempts: config.MaxRetryAttempts,
-		DeliveryTimeout:  config.DeliveryTimeout,
-		OrderedEvent:     config.OrderedEvent,
+		RateLimit:         config.RateLimit,
+		MaxRetryAttempts:  config.MaxRetryAttempts,
+		DeliveryTimeout:   config.DeliveryTimeout,
+		DisableDeadLetter: config.DisableDeadLetter,
+		OrderedEvent:      config.OrderedEvent,
 	}
 	switch config.OffsetType {
 	case pb.SubscriptionConfig_LATEST:
@@ -225,10 +226,11 @@ func fromPbSubscriptionConfig(config *pb.SubscriptionConfig) primitive.Subscript
 
 func toPbSubscriptionConfig(config primitive.SubscriptionConfig) *pb.SubscriptionConfig {
 	to := &pb.SubscriptionConfig{
-		RateLimit:        config.RateLimit,
-		MaxRetryAttempts: config.MaxRetryAttempts,
-		DeliveryTimeout:  config.DeliveryTimeout,
-		OrderedEvent:     config.OrderedEvent,
+		RateLimit:         config.RateLimit,
+		MaxRetryAttempts:  config.MaxRetryAttempts,
+		DeliveryTimeout:   config.DeliveryTimeout,
+		DisableDeadLetter: config.DisableDeadLetter,
+		OrderedEvent:      config.OrderedEvent,
 	}
 	switch config.OffsetType {
 	case primitive.LatestOffset:
