+++ internal/trigger/worker.go
@@ -254,7 +254,8 @@ func (w *worker) getTriggerOptions(subscription *primitive.Subscription) []trigg
 	opts = append(opts, trigger.WithRateLimit(config.RateLimit),
 		trigger.WithDeliveryTimeout(config.DeliveryTimeout),
 		trigger.WithMaxRetryAttempts(config.GetMaxRetryAttempts()),
-		trigger.WithEventbus(subscription.EventBus),
+		trigger.WithRetry(subscription.EventBus),
+		trigger.WithDeadLetter(subscription.EventBus, config.DisableDeadLetter),
 		trigger.WithOrdered(config.OrderedEvent),
 		trigger.WithGoroutineSize(w.config.SendEventGoroutineSize),
 		trigger.WithSendBatchSize(w.config.SendEventBatchSize),
