+++ proto/proto/meta.proto
@@ -167,7 +167,8 @@ message SubscriptionConfig {
   // delivery timeout, unit milliseconds
   uint32 delivery_timeout = 4;
   optional uint32 max_retry_attempts = 5;
-  bool ordered_event = 6;
+  bool disable_dead_letter = 6;
+  bool ordered_event = 7;
 }
 
 message Filter {
